requirements.txt
jupyter==1.0.0
matplotlib==3.0.2
mypy==0.641
numpy==1.15.4
pandas==0.23.4
scipy==1.1.0
scikit-learn==0.20.1
pytest==4.0.1
tqdm==4.28.1
pyyaml
GPy
tangent
flask
psutil
ipdb
